fix(running-script): add warning when a script does not exist in a package instead of crashing
