+++ src/scripts.ts
@@ -1,7 +1,39 @@
 import { getStagedFilesDirectory } from './git'
-import { getPackageInfo, execPackageManagerScript } from './utils'
+import { getPackageInfo, execPackageManagerScript, PackageJSON } from './utils'
 import { getConfig } from './config'
-import { handleScriptRunLog } from './logger'
+import ora from 'ora'
+import { logSuccess, logWarning } from './logger'
+import locale from './locale'
+
+const handleScriptRun = async (
+  scriptRunFunction: () => Promise<string>,
+  script: string,
+  packageJSON: PackageJSON
+): Promise<void> => {
+  const spinner = ora()
+  const packageName = packageJSON?.name || 'Unknown'
+
+  if (!packageJSON?.scripts || !packageJSON.scripts[script]) {
+    logWarning(locale.warning.scriptNotFound(script, packageName))
+    spinner.clear()
+    return
+  }
+
+  try {
+    spinner.start(locale.info.runningScript(script, packageName))
+
+    const scriptOutput = await scriptRunFunction()
+
+    console.log(scriptOutput)
+    spinner.succeed(locale.success.scriptRun(script, packageName))
+
+    logSuccess(locale.success.scriptRun(script, packageName))
+  } catch (error) {
+    console.error(error.message)
+    spinner.fail(locale.errors.scriptFailed(script, packageName))
+    process.exit(1)
+  }
+}
 
 export const runCommands = async (): Promise<void> => {
   const configuration = getConfig()
@@ -10,10 +42,11 @@ export const runCommands = async (): Promise<void> => {
 
   for (const pkg of Object.values(packagesInfo)) {
     for (const script of configuration.scripts) {
-      const runPackageScript = (): Promise<string> =>
-        execPackageManagerScript(configuration.packageManager, script, pkg.directory)
+      const runPackageScript = (): Promise<string> => {
+        return execPackageManagerScript(configuration.packageManager, script, pkg.directory)
+      }
 
-      await handleScriptRunLog(runPackageScript, script, pkg.json.name)
+      await handleScriptRun(runPackageScript, script, pkg.json)
     }
   }
 }
